test_scripts/requirements_old.txt
aniso8601==9.0.1
argcomplete==1.10.0
autopep8==1.5.6
beautifulsoup4==4.8.0
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cymem==2.0.5
docx2txt==0.8
EbookLib==0.17.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
extract-msg==0.23.1
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
IMAPClient==2.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lxml==4.6.3
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.2
olefile==0.46
packaging==20.9
pandas==1.2.4
pathy==0.5.2
pdfminer.six==20200517
pdfplumber==0.5.27
phonenumbers==8.12.21
Pillow==8.2.0
plac==1.1.3
preshed==3.0.5
pycodestyle==2.7.0
pycryptodome==3.10.1
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2021.1
regex==2021.4.4
requests==2.25.1
shortuuid==1.0.1
six==1.12.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
SpeechRecognition==3.8.1
srsly==2.4.1
stemming==1.0.1
textract==1.6.3
thinc==8.0.3
tika==1.24
toml==0.10.2
tqdm==4.60.0
typer==0.3.2
tzlocal==1.5.1
urllib3==1.26.4
Wand==0.6.6
wasabi==0.8.2
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.4.0
